Move all scraping functions to utils.scrape_utils
